pyproject.toml
[project]
name = "DocumentIngestionPipeline-AllCognix"
version = "0.1.0"
description = "Scalable document ingestion pipeline using FastAPI, DBOS, and LlamaIndex"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbos>=2.7.0",
    "fastapi[standard]>=0.128.0",
    "ipykernel>=7.1.0",
    "llama-index>=0.14.12",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-readers-docling>=0.4.2",
    "llama-index-vector-stores-postgres>=0.7.2",
    "onnxruntime>=1.23.2",
    "openai-whisper>=20250625",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "url-parser>=3.0.3",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
filter_files = true
atomic = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
